Conflicts:
	deps/ucdata/ucdata.cpp
	nui3.xcodeproj/project.pbxproj
	src/File/nglPath.cpp

+++ src/Text/nuiLabel.cpp
@@ -164,6 +164,12 @@ void nuiLabel::InitAttributes()
                (nglString(_T("VMargin")), nuiUnitNone,
                 nuiMakeDelegate(this, &nuiLabel::GetVMargin), 
                 nuiMakeDelegate(this, &nuiLabel::SetVMargin)));
+
+  AddAttribute(new nuiAttribute<bool>
+               (nglString(_T("UseEllipsis")), nuiUnitYesNo,
+                nuiMakeDelegate(this, &nuiLabel::GetUseEllipsis), 
+                nuiMakeDelegate(this, &nuiLabel::UseEllipsis)));
+  
 }
 
 void nuiLabel::InitProperties()
@@ -401,12 +407,17 @@ nuiRect nuiLabel::CalcIdealSize()
 
 bool nuiLabel::SetRect(const nuiRect& rRect)
 {
+  bool needRecalcLayout = false;
+
+  if (mUseEllipsis || mWrapping)
+    needRecalcLayout = (rRect.GetWidth() != mRect.GetWidth());
+    
   nuiWidget::SetRect(rRect);
 
   nuiRect ideal(mIdealLayoutRect);
 
 
-  if (ideal.GetWidth() > mRect.GetWidth())
+  if (needRecalcLayout || ideal.GetWidth() > mRect.GetWidth())
   {
     if (mUseEllipsis)
     {
@@ -414,9 +425,12 @@ bool nuiLabel::SetRect(const nuiRect& rRect)
       nuiSize diff = ideal.GetWidth() - mRect.GetWidth();
       int NbLetterToRemove = ToNearest(diff / (ideal.GetWidth() / mText.GetLength())) + 3;
       nglString text = mText;
-      int len = text.GetLength();
-      text.DeleteRight(MIN(NbLetterToRemove, len));
-      text.Append(_T("..."));
+      if (NbLetterToRemove > 0)
+      {
+        int len = text.GetLength();
+        text.DeleteRight(MIN(NbLetterToRemove, len));
+        text.Append(_T("..."));
+      }
       delete mpLayout;
       mpLayout = new nuiTextLayout(mpFont);
       mpLayout->SetWrapX(0);
@@ -640,6 +654,10 @@ void nuiLabel::UseEllipsis(bool useEllipsis)
 {
   mUseEllipsis = useEllipsis;
   InvalidateLayout();
+}
+bool nuiLabel::GetUseEllipsis() const
+{
+  return mUseEllipsis;
 }
 
 void nuiLabel::SetOrientation(nuiOrientation Orientation)
